requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.10
async-timeout==5.0.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
coverage==7.8.2
cron-descriptor==1.4.5
dill==0.4.0
dj-config-url==0.1.1
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-api-logger==1.1.16
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
kombu==5.5.0
mccabe==0.7.0
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycodestyle==2.13.0
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.7
PySocks==1.7.1
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.6
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
selenium==4.23.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
